make some macros DependencyControl-compatible
