+++ macros/lyger.CircleText.lua
@@ -22,23 +22,18 @@ Only works on static lines. If you want the line to move or rotate, use another 
 
 ]]--
 
-script_name="Circular text"
-script_description="Puts the text on a circular arc centered on the origin"
-script_version="0.1"
+script_name = "Circular text"
+script_description = "Puts the text on a circular arc centered on the origin."
+script_version = "0.1.0"
+script_author = "lyger"
+script_namespace = "lyger.CircleText"
 
-include("karaskel.lua")
---require "re"
-
---Creates a deep copy of the given table
-local function deep_copy(source_table)
-	new_table={}
-	for key,value in pairs(source_table) do
-		--Let's hope the recursion doesn't break things
-		if type(value)=="table" then value=deep_copy(value) end
-		new_table[key]=value
-	end
-	return new_table
-end
+local DependencyControl = require("l0.DependencyControl")
+local rec = DependencyControl{
+    feed = "https://raw.githubusercontent.com/TypesettingTools/lyger-Aegisub-Scripts/master/DependencyControl.json",
+    {"aegisub.util", "karaskel"}
+}
+local util = rec:requireModules()
 
 --[[
 Tags that can have any character after the tag declaration:
@@ -166,77 +161,77 @@ end
 
 --Main processing function
 function circle_text(sub,sel)
-	
+
 	--Read in styles and meta
 	local meta,styles = karaskel.collect_head(sub, false)
-	
+
 	for si,li in ipairs(sel) do
-		
+
 		--Progress report
 		aegisub.progress.task("Processing line "..si.."/"..#sel)
 		aegisub.progress.set(100*si/#sel)
-		
+
 		--Read in the line
 		line=sub[li]
-		
+
 		--Preprocess
 		karaskel.preproc_line(sub,meta,styles,line)
-		
+
 		--Get position and origin
 		px,py=get_pos(line)
 		ox,oy=get_org(line)
-		
+
 		--Make sure pos and org are not the same
 		if px==ox and py==oy then
 			aegisub.log(1,"Error on line %d: Position and origin cannot be the same!",li)
 			return
 		end
-		
+
 		--Get radius
 		radius=distance(px,py,ox,oy)
-		
+
 		--Remove \pos and \move
 		--If your line was non-static, too bad
 		line.text=line_exclude(line.text,{"pos","move"})
-		
+
 		--Make sure line starts with a tag block
 		if line.text:find("^{")==nil then
 			line.text="{}"..line.text
 		end
-		
+
 		--Rotation direction: positive if each character adds to the angle,
 		--negative if each character subtracts from the angle
 		rot_dir=sign(py-oy)
-		
+
 		--Add the \pos back with recalculated position
 		line.text=line.text:gsub("^{",string.format("{\\pos(%d,%d)",ox,oy+rot_dir*radius))
-		
+
 		--Get z rotation
 		--Will only take the first one, because if you wanted the text to be on a circular arc,
 		--why do you have more than one z rotation tag in the first place?
 		_,_,zrot=line.text:find("\\frz([%-%.%d]+)")
 		zrot=zrot or line.styleref.angle
-		
+
 		--Make line table
 		line_table={}
 		for thistag,thistext in line.text:gmatch("({[^{}]*})([^{}]*)") do
 			table.insert(line_table,{tag=thistag,text=thistext})
 		end
-		
+
 		--Where data on the character widths will be stored
 		char_data={}
-		
+
 		--Total width of line
 		cum_width=0
-		
+
 		--Stores current state of the line as style table
-		current_style=deep_copy(line.styleref)
-		
+		current_style = util.deep_copy(line.styleref)
+
 		--First pass to collect data on character widths
 		for i,val in ipairs(line_table) do
-			
+
 			char_data[i]={}
-			
+
 			--Fix style tables to reflect override tags
 			local _,_,font_name=val.tag:find("\\fn([^\\{}]+)")
 			local _,_,font_size=val.tag:find("\\fs([%-%.%d]+)")
@@ -245,7 +240,7 @@ function circle_text(sub,sel)
 			local _,_,font_sp=val.tag:find("\\fsp([%-%.%d]+)")
 			local _,_,_bold=val.tag:find("\\b([01])")
 			local _,_,_italic=val.tag:find("\\i([01])")
-			
+
 			current_style.fontname=font_name or current_style.fontname
 			current_style.fontsize=tonumber(font_size) or current_style.fontsize
 			current_style.scale_x=tonumber(font_scx) or current_style.scale_x
@@ -259,31 +254,31 @@ function circle_text(sub,sel)
 				if _italic=="1" then current_style.italic=true
 				else current_style.italic=false end
 			end
-			
-			val.style=deep_copy(current_style)
-			
+
+			val.style = util.deep_copy(current_style)
+
 			--Collect width data on each char
 			for thischar in val.text:gmatch(".") do
 				cwidth=aegisub.text_extents(val.style,thischar)
 				table.insert(char_data[i],{char=thischar,width=cwidth})
 			end
-			
+
 			--Increment cumulative width
 			cum_width=cum_width+aegisub.text_extents(val.style,val.text)
-			
+
 		end
-		
+
 		--The angle that the rotation will begin at
 		start_angle=zrot-(rot_dir*arc_angle(cum_width,radius))/2
-		
+
 		rebuilt_text=""
 		cum_rot=0
-		
+
 		--Second pass to rebuild line with new tags
 		for i,val in ipairs(line_table) do
-		
+
 			rebuilt_text=rebuilt_text..val.tag:gsub("\\fsp[%-%.%d]+",""):gsub("\\frz[%-%.%d]+","")
-			
+
 			for k,tchar in ipairs(char_data[i]) do
 				--Character spacing should be the average of this character's width and the next one's
 				--For spacing, scale width back up by the character's relevant scale_x,
@@ -296,24 +291,24 @@ function circle_text(sub,sel)
 				else
 					this_width=i~=#line_table and (tchar.width+char_data[i+1][1].width)/2 or 0
 					this_spacing=i~=#line_table
-						and -1*((tchar.width*100/val.style.scale_x 
+						and -1*((tchar.width*100/val.style.scale_x
 							+ char_data[i+1][1].width*100/line_table[i+1].style.scale_x)/2
 							-val.style.spacing)
 						or 0
 				end
-				
+
 				rebuilt_text=rebuilt_text..string.format("{\\frz%.3f\\fsp%.2f}%s",
 					(start_angle+rot_dir*cum_rot)%360,this_spacing,tchar.char)
-				
+
 				cum_rot=cum_rot+arc_angle(this_width,radius)
 			end
 		end
-		
+
 		--[[
 		--Fuck the re library. Maybe I'll come back to this
 		whitespaces=re.find(rebuilt_text,
 			'(\{\\\\frz[\\d\\.\\-]+\\\\fsp[\\d\\.\\-]+\}\\S)((?:\{\\\\frz[\\d\\.\\-]+\\\\fsp[\\d\\.\\-]+\}\\s)+)')
-		
+
 		for j=1,#whitespaces-1,2 do
 			first_tag=whitespaces[j].str
 			other_tags=whitespaces[j+1].str
@@ -329,15 +324,15 @@ function circle_text(sub,sel)
 			rebuilt_text=rebuilt_text:gsub(first_tag..other_tags,
 				first_tag:gsub("\\fsp[%d%.%-]+",string.format("\\fsp%.2f",total_space))..string.rep(" ",total_wsp))
 		end]]--
-		
+
 		line.text=rebuilt_text:gsub("}{","")
-		
+
 		sub[li]=line
-		
+
 	end
-	
+
 	aegisub.set_undo_point(script_name)
-	
+
 end
 
-aegisub.register_macro(script_name,script_description,circle_text)
+rec:registerMacro(circle_text)
