+++ macros/lyger.KaraReplacer.lua
@@ -1,33 +1,40 @@
-script_name="Karaoke replacer"
-script_description="Replaces the syllables of a verse"
-script_version="0.2.1"
+script_name = "Karaoke replacer"
+script_description = "Replaces the syllables of a verse."
+script_version = "0.2.1"
+script_author = "lyger"
+script_namespace = "lyger.KaraReplacer"
+
+local DependencyControl = require("l0.DependencyControl")
+local rec = DependencyControl{
+    feed = "https://raw.githubusercontent.com/TypesettingTools/lyger-Aegisub-Scripts/master/DependencyControl.json"
+}
 
 --Fuck it, I should comment this code. Her goes
 function kara_replace(sub,sel)
 	for si,li in ipairs(sel) do
 		--Read in line
-		line=sub[li]
-		
+		local line = sub[li]
+
 		--Split at karaoke tags and create table of tags and text
-		line_table={}
-		
+		local line_table = {}
+
 		for tag,text in line.text:gmatch("({[^{}]*\\[kK][^{}]*})([^{}]*)") do
 			table.insert(line_table,{["tag"]=tag,["text"]=text})
 		end
-		
+
 		--Put the line back together with spaces between syllables
-		rebuilt_original=""
-		
+		local rebuilt_parts = {}
+
 		for i,val in ipairs(line_table) do
-			rebuilt_original=rebuilt_original..val.text.." "
+			rebuilt_parts[i] = val.text
 		end
-		
+
 		--Add some padding so it displays better
-		rebuilt_original=rebuilt_original..string.rep(" ",math.floor(rebuilt_original:len()/2))
-		
+		local rebuilt_original = table.concat(rebuilt_parts, " ")
+		rebuilt_original = rebuilt_original .. string.rep(" ", math.floor(rebuilt_original:len()/2) -1)
+
 		--Dialog display
-		config=
-		{
+		local config = {
 			{
 				class="label",
 				label=rebuilt_original,
@@ -40,10 +47,9 @@ function kara_replace(sub,sel)
 				x=0,y=1,width=1,height=1
 			}
 		}
-		
+
 		--Instructions
-		help_config=
-		{
+		local help_config = {
 			{
 				class="label",
 				label=
@@ -62,92 +68,90 @@ function kara_replace(sub,sel)
 				x=0,y=0,width=1,height=1
 			}
 		}
-		
+
 		--Show dialog and get input for each line
-		pressed=nil
-		
+		local pressed
+
 		repeat
 			pressed,result=aegisub.dialog.display(config,{"Next line","Help"})
 			if pressed=="Help" then
 				aegisub.dialog.display(help_config,{"OK"})
 			end
 		until pressed~="Help"
-		
+
 		--Split input at spaces and store in table
-		replace={}
-		
+		local replace = {}
+
 		for newsyl in result["replace"]:gsub("  ","\t "):gmatch("[^ ]+") do
 			newsyl=newsyl:gsub("\t"," ")
 			table.insert(replace,newsyl)
 		end
-		
-		rebuilt_text=""
-		
+
+		local rebuilt_text, r = {}, 1
 		--Indices of original and replacement tables
-		oi=1
-		ri=1
-		
+		local oi, ri = 1, 1
+
 		while oi<=#line_table do
 			--Skip if it's a blank syl (used for padding) or we're out of replacements
 			if line_table[oi].text:len()>0 and replace[ri]~=nil then
 				--Handle splitting syls
 				if replace[ri]:find("|")~=nil then
-					
+
 					--Split the replacement line at | characters
 					subtab={}
 					for subsyl in replace[ri]:gmatch("[^|]+") do
 						table.insert(subtab,subsyl)
 					end
-					
+
 					--Find the original time of the karaoke syllable
-					otime=tonumber(line_table[oi].tag:match("\\[kK][fo]?(%d+)"))
+					local otime = tonumber(line_table[oi].tag:match("\\[kK][fo]?(%d+)"))
 					--The remaining time (for last syl, to ensure they add up to the original time)
-					ltime=otime
+					local ltime = otime
 					--Add all but the last syl
 					for x=1,#subtab-1,1 do
 						--To minimize truncation error, alternate between ceil and floor
-						ttime=0
+						local ttime = 0
 						if x%2==1 then
-							ttime=math.floor(otime/#subtab)
+							ttime = math.floor(otime/#subtab)
 						else
-							ttime=math.ceil(otime/#subtab)
+							ttime = math.ceil(otime/#subtab)
 						end
-						rebuilt_text=rebuilt_text..
-							line_table[oi].tag:gsub("(\\[kK][fo]?)%d+","\1"..tostring(ttime))..
-							subtab[x]
+						rebuilt_text[r] = line_table[oi].tag:gsub("(\\[kK][fo]?)%d+","\1"..tostring(ttime))
+						rebuilt_text[r+1], r = subtab[x], r+2
 						ltime=ltime-ttime
 					end
 					--Add the last syl
-					rebuilt_text=rebuilt_text..
-						line_table[oi].tag:gsub("(\\[kK][fo]?)%d+","\1"..tostring(ltime))..
-						subtab[#subtab]
-				
+					rebuilt_text[r] = line_table[oi].tag:gsub("(\\[kK][fo]?)%d+","\1"..tostring(ltime))
+					rebuilt_text[r+1], r = subtab[#subtab], r+2
+
 				--Handle merging syls
 				--Only merge if it's not the last syl
 				elseif replace[ri]:find("+")~=nil and oi<#line_table then
-					temp_tag=line_table[oi].tag
+					local temp_tag = line_table[oi].tag
 					oi=oi+1
 					stime=tonumber(line_table[oi].tag:match("\\[kK][fo]?(%d+)"))
 					temp_tag=temp_tag:gsub("(\\[kK][fo]?)(%d+)",function(a,b)
 							return a..tostring(tonumber(b)+stime)
 						end)
-					rebuilt_text=rebuilt_text..temp_tag..replace[ri]:gsub("+","")
-					
+					rebuilt_text[r], rebuilt_text[r+1] = temp_tag, replace[ri]:gsub("+","")
+					r = r+2
+
 				--The usual replacement
 				else
-					rebuilt_text=rebuilt_text..line_table[oi].tag..replace[ri]:gsub("_","")
+					rebuilt_text[r], rebuilt_text[r+1] = line_table[oi].tag, replace[ri]:gsub("_","")
+					r = r+2
 				end
-				
+
 				--Increment indices
 				oi=oi+1
 				ri=ri+1
 			else
-				rebuilt_text=rebuilt_text..line_table[oi].tag
+				rebuilt_text[r], r = line_table[oi].tag, r+1
 				oi=oi+1
 			end
 		end
-		
-		line.text=rebuilt_text
+
+		line.text = table.concat(rebuilt_text)
 		sub[li]=line
 		if finished then break end
 	end
@@ -159,19 +163,19 @@ end
 function kara_replace_old(sub,sel)
 	for si,li in ipairs(sel) do
 		line=sub[li]
-		
+
 		line_table={}
-		
+
 		for tag,text in line.text:gmatch("({[^{}]*\\[kK][^{}]*})([^{}]*)") do
 			table.insert(line_table,{["tag"]=tag,["text"]=text})
 		end
-		
+
 		rebuilt_text=""
-		
+
 		finished=false
-		
+
 		for i,val in ipairs(line_table) do
-			
+
 			local function hl_syl(lt,idx)
 				result=""
 				for k,a in ipairs(lt) do
@@ -180,7 +184,7 @@ function kara_replace_old(sub,sel)
 				end
 				return result
 			end
-			
+
 			if val.text:len()<1 or finished then
 				rebuilt_text=rebuilt_text..val.tag..val.text
 			else
@@ -217,4 +221,4 @@ function kara_replace_old(sub,sel)
 	end
 end
 
-aegisub.register_macro(script_name,script_description,kara_replace)
+rec:registerMacro(kara_replace)
